lod blending based on time instead of frames

+++ browser/src/vts-libbrowser/camera.hpp
@@ -84,7 +84,7 @@ class OldDraw
 public:
     TileId trav;
     TileId orig;
-    uint32 age;
+    double age;
 
     OldDraw(const CurrentDraw &current);
     OldDraw(const TileId &id);
